+++ saleor/tests/e2e/account/utils/user.py
@@ -5,6 +5,7 @@ USER_QUERY = (
     """
     query User($id: ID!) {
         user(id: $id) {
+            id
             email
             firstName
             lastName