+++ saleor/tests/e2e/orders/test_order_keep_address_information.py
@@ -1,10 +1,9 @@
 import pytest
 
 from .. import ADDRESS_DE, DEFAULT_ADDRESS
-from ..account.utils import account_address_delete, get_user
+from ..account.utils import account_address_delete, create_customer, get_user
 from ..product.utils.preparing_product import prepare_product
 from ..shop.utils.preparing_shop import prepare_default_shop
-from ..users.utils import create_customer
 from ..utils import assert_address_data, assign_permissions
 from .utils import (
     draft_order_complete,